refactor: Move SyncJobsActor to a separate file
